Moved all the root directory 2024 config files into 2024 sub-directory, for consistency. (#234)
