+++ src/components/2023/BecomeSponsor/index.vue
@@ -18,7 +18,7 @@
 	</section>
 </template>
 <script>
-import { getButton } from "@/config/buttons.js";
+import { getButton } from "@/config/2023/buttons.js";
 export default {
 	data() {
 		return {
