+++ src/components/2024/BecomeSponsor/index.vue
@@ -15,11 +15,12 @@
 			>
 				<button>BECOME A SPONSOR</button>
 			</a>
+			
 		</div>
 	</section>
 </template>
 <script>
-import { getButton } from "@/config/buttons.js";
+import { getButton } from "@/config/2024/buttons.js";
 export default {
 	data() {
 		return {
