+++ src/config/2024/photos.json
