+++ src/config/2024/speakers.json
